offline check

+++ src/actions/domOperation.js
@@ -1,11 +1,15 @@
+
 import { apiCall } from "../apiCalls/taskAPI.js";
 import { createTask } from "../components/task.js";
 
 const url = "https://todo-list-app-backend-7.herokuapp.com/tasks";
+// const url = "http://127.0.0.1:3000/tasks";
+
+
 
 export const getTodos = async () => {
      let allTasks = await apiCall(`${url}`);
-     console.log(allTasks);
+    //  console.log(allTasks);
     allTasks.forEach((item) => {
         createTask(item);
        
@@ -15,10 +19,15 @@ export const getTodos = async () => {
 
 
 export const addTodo = async (event) => {
+   
     event.preventDefault();
     let taskField = document.getElementById("taskField");
     let inputValue = taskField.value;
     taskField.value = "";
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     let taskObj = {
         method: 'POST',
         body: JSON.stringify({content: inputValue, 
@@ -28,17 +37,25 @@ export const addTodo = async (event) => {
             "Content-Type": "application/json"
         } 
     }
-    const addTask = await apiCall(`${url}`, taskObj);
-    createTask(addTask);
+   
+        const addTask = await apiCall(`${url}`, taskObj);
+        createTask(addTask);
+  
+    console.log(addTask);
 }
 
 
 export const deleteTodo = (e) => {
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     let taskObj = {
         method: 'DELETE'
     }
+   
     const id = e.target.parentElement.id;
-    console.log(id);
+    // console.log(id);
     const deleteTask = document.getElementById(id);
     apiCall(`${url}/${id}`, taskObj);
     deleteTask.remove();
@@ -48,15 +65,25 @@ export const deleteTodo = (e) => {
 
 
 export const updateTodo = (e) => {
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     const enableInput = e.target.parentElement.childNodes[1];
-    enableInput.disabled = false;
-    e.target.parentElement.childNodes[3].style.display = 'none';
-    e.target.parentElement.childNodes[4].style.display = 'unset';
-
-
+ 
+        enableInput.disabled = false;
+        e.target.parentElement.childNodes[3].style.display = 'none';
+        e.target.parentElement.childNodes[4].style.display = 'unset';   
+    
+    
 }
 
+
 export const doneEditTask = async (e) => {
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     const enableInput = e.target.parentElement.childNodes[1];
     const parentelem = e.target.parentElement;
     const id = e.target.parentElement.id;
